+++ test/io_tests/request_response_tests/disconnect_test.py
@@ -2,6 +2,9 @@
 
 from unittest.mock import patch
 
+import pytest
+
+from xknx.exceptions import RequestResponseError
 from xknx.io.request_response import Disconnect
 from xknx.io.transport import UDPTransport
 from xknx.knxip import (
@@ -41,7 +44,8 @@ class TestDisconnect:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await disconnect.start()
+            with pytest.raises(RequestResponseError):
+                await disconnect.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
@@ -54,19 +58,14 @@ class TestDisconnect:
         err_knxipframe = KNXIPFrame.init_from_body(
             DisconnectResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            disconnect.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(disconnect).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        disconnect.response_rec_callback(err_knxipframe, HPAI(), None)
+        assert disconnect._response is None
+        assert disconnect._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(DisconnectResponse())
         disconnect.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert disconnect.response is not None
+        assert disconnect._response is not None
 
     async def test_disconnect_route_back_true(self) -> None:
         """Test disconnecting from KNX bus."""
@@ -92,7 +91,8 @@ class TestDisconnect:
             patch("xknx.io.transport.UDPTransport.getsockname") as mock_udp_getsockname,
         ):
             mock_udp_getsockname.return_value = ("192.168.1.3", 4321)
-            await disconnect.start()
+            with pytest.raises(RequestResponseError):
+                await disconnect.request()
             mock_udp_send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
@@ -105,16 +105,11 @@ class TestDisconnect:
         err_knxipframe = KNXIPFrame.init_from_body(
             DisconnectResponse(status_code=ErrorCode.E_CONNECTION_ID)
         )
-        with patch("logging.Logger.debug") as mock_warning:
-            disconnect.response_rec_callback(err_knxipframe, HPAI(), None)
-            mock_warning.assert_called_with(
-                "Error: KNX bus responded to request of type '%s' with error in '%s': %s",
-                type(disconnect).__name__,
-                type(err_knxipframe.body).__name__,
-                ErrorCode.E_CONNECTION_ID,
-            )
+        disconnect.response_rec_callback(err_knxipframe, HPAI(), None)
+        assert disconnect._response is None
+        assert disconnect._error_code is ErrorCode.E_CONNECTION_ID
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(DisconnectResponse())
         disconnect.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert disconnect.response is not None
+        assert disconnect._response is not None